feat: Implement SOCKS5 client proxy
Implement (optional) socks5 proxy in the iox client library.

This allows a local IOx client (e.g. a development version of conductor) to talk to IOx servers
running in a local k8s cluster even if the pod networking is not directly routable.

Among other things, this offers a simpler way to develop on a mac and test conductor locally
without having to cross-compile it for linux and run it in a container (or using complicated things
like teleport).

Demo:

```console
$ kubectl krew install socks5-proxy
...
$ kubectl socks5-proxy &
...
$ ALL_PROXY=socks5://127.0.0.1:1080 ./target/debug/influxdb_iox -h http://10.1.0.7:8082 server get
132072
```